refactor(progress): centralize streaming progress formatting in helpers
